agregue sweetalert para el borrado de favoritos

+++ src/store/favoritos.js
@@ -1,5 +1,6 @@
 import {defineStore} from 'pinia';
 import { ref } from 'vue';
+import Swal from 'sweetalert2';
 
 export const useFavoritosStore = defineStore('favoritos', () =>{
     const favoritos = ref([]);
@@ -13,9 +14,27 @@ if(localStorage.getItem("favorito"))
         localStorage.setItem('favorito', JSON.stringify(favoritos.value))
     } 
 
-    const deleteFavorito = (id) =>{
-        favoritos.value = favoritos.value.filter(item => item.id !== id)
-        localStorage.setItem('favorito', JSON.stringify(favoritos.value))
+    const deleteFavorito = (id) =>{       
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+              favoritos.value = favoritos.value.filter(item => item.id !== id)
+              localStorage.setItem('favorito', JSON.stringify(favoritos.value))
+              Swal.fire(
+                'Deleted!',
+                'The pokemon has been deleted of yours favorite.',
+                'success'
+              )
+            }
+          })
     }
 
     const findByName = (nombre) => {
